Fixed a regression issue where excessive memory was allocated for the GLVolumes before sending to the GPU driver. The following commits were partially reverted:
304e31f87c06bb9b1f8d0d2a44efa4c1df379d86 Removed GLVolume non-VBO rendering
be380b5bce2ed53b7f224cb77057aa4e8328ea70 GLVolume and GLIndexedVertexArray refactored to send data to gpu at the first render call

Namely, the GLVolume buffers are "shrink to size"'d before sending their
content to the OpenGL driver, and the vertex buffers are populated
as quickly as possible from the GLVolume, so that the same buffer is not
kept twice in RAM on systems, where the RAM is shared with the graphics
card.

Also the memory allocation reporting was improved for the GLVolumes.