Initial add of Creality Sermoon V1 Printer and 3 process configs (#4956)
* Add files via upload

* Add files via upload

* Add files via upload

* Add files via upload

* Add files via upload

* Update Creality Sermoon V1 0.4 nozzle.json

Changed the default process for the Sermoon V1 Printer.

* Update 0.20mm Standard @Creality Sermoon V1.json

Reduced speeds of .20mm process for Sermoon V1 to no more than 195mm/s as not all printers which may use this are running Klipper.  It is set to Marlin for the flavour but is targeted at a Klippered printer.

* Update 0.16mm Optimal @Creality Sermoon V1.json

Remove Arc Fitting and reduce elephant foot.

* Update 0.20mm Standard @Creality Sermoon V1.json

Remove Arc Fitting and reduce elephant foot.

* Update 0.28mm Standard @Creality Sermoon V1.json

Remove Arc Fitting and reduce elephant foot.